requirements.txt
Flask
Flask-SQLAlchemy
requests
Flask-Migrate
python-dotenv
gunicorn
